requirements.txt
pyserial
pyqt5
pyqtgraph
numpy